tests/wpt/tests/mathml/relations/html5-tree/a-type-setter.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathMLAnchorElement.type setter</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <math>
      <a id="test" href="a">Example Link</a>
    </math>
  </body>
  <script>
      test(function() {
          var a = document.getElementById("test");
          a.type = "text/plain"
          assert_equals(a.type, "text/plain");
          assert_equals(a.getAttribute('type'), "text/plain");
      }, "Test anchor's type setter");
  </script>
</html>
