third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/a-type-setter-expected.txt

This is a testharness.js-based test.
[FAIL] Test anchor's type setter
  assert_equals: expected (string) "text/plain" but got (object) null
Harness: the test ran to completion.
